requirements.txt
streamlit
pandas
requests
beautifulsoup4
spotipy
python-dotenv